home: separate state from topbar
