+++ components/src/docs/docs.js
@@ -20,7 +20,8 @@ import createStore from './store';
 
 const http = new Web3.providers.HttpProvider(process.env.RPC_ADDRESS || '/rpc/');
 const web3 = new Web3(http);
-web3._extend(web3extensions(web3));
+web3extensions(web3)
+  .map(extension => rawWeb3._extend(extension));
 
 const store = createStore(middlewares());
