+++ components/src/util/web3.extensions.js
@@ -1,7 +1,24 @@
 export default function web3extensions (web3) {
-  const { Method } = web3._extend;
+  const { Method, formatters } = web3._extend;
 
-  return {
+  return [{
+    property: 'personal',
+    methods: [
+      new Method({
+        name: 'signAndSendTransaction',
+        call: 'personal_signAndSendTransaction',
+        params: 2,
+        inputFormatter: [formatters.inputTransactionFormatter, null]
+      }),
+      new Method({
+        name: 'signerEnabled',
+        call: 'personal_signerEnabled',
+        params: 0,
+        inputFormatter: []
+      })
+    ],
+    properties: []
+  }, {
     property: 'ethcore',
     methods: [
       new Method({
@@ -15,8 +32,14 @@ export default function web3extensions (web3) {
         call: 'ethcore_gasPriceStatistics',
         params: 0,
         outputFormatter: a => a.map(web3.toBigNumber)
+      }),
+      new Method({
+        name: 'unsignedTransactionsCount',
+        call: 'ethcore_unsignedTransactionsCount',
+        params: 0,
+        inputFormatter: []
       })
     ],
     properties: []
-  };
+  }];
 }
