+++ home/web/client/components/Root/Root.css
@@ -1,0 +1,7 @@
+.toasts {
+  position: fixed;
+  top: 30px;
+  right: 30px;
+  z-index: 100000;
+  width: 200px;
+}
