+++ home/web/client/utils/accounts.js
@@ -1,0 +1,7 @@
+export function fixAccountNames (names, accounts) {
+  const copy = Object.assign({}, names);
+  return accounts.reduce((memo, acc, idx) => {
+    memo[acc] = names[acc] || `Account ${idx + 1}`;
+    return memo;
+  }, copy);
+}
