+++ home/web/client/utils/extension.js
@@ -1,0 +1,5 @@
+import { CHROME_EXT_ID } from '../constants/extension';
+
+export function fetchIsExtensionInstalled (cb) {
+  window.chrome.runtime.sendMessage(CHROME_EXT_ID, 'version', cb);
+}
