+++ home/web/client/utils/signer.js
@@ -1,0 +1,3 @@
+export function signerUrl (port) {
+  return `http://127.0.0.1:${port}/index.html`;
+}
